+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Add support for baggage propagation (#51)
 - Updating dependencies to 0.14b0 (#51)
-- Add support for tracecontext propagation (#53)
+- Add support for tracecontext propagation (#52)
 
 ## 0.13b0